Promote ChiselException to package chisel3
It was previously aliased from chisel3.internal which is bad for
ScalaDoc.
