+++ core/src/main/scala/chisel3/RawModule.scala
@@ -141,6 +141,9 @@ package object internal {
   @deprecated("This function has moved to chisel3.experimental", "Chisel 3.6")
   val noPrefix = chisel3.experimental.noPrefix
 
+  @deprecated("This function has moved to chisel3", "Chisel 3.6")
+  type ChiselException = chisel3.ChiselException
+
   import scala.annotation.implicitNotFound
   @implicitNotFound("You are trying to access a macro-only API. Please use the @public annotation instead.")
   trait MacroGenerated
