+++ core/src/main/scala/chisel3/experimental/hierarchy/core/Lookupable.scala
@@ -10,15 +10,7 @@ import scala.collection.mutable.HashMap
 import chisel3._
 import chisel3.experimental.dataview.{isView, reify, reifySingleData}
 import chisel3.internal.firrtl.{Arg, ILit, Index, ModuleIO, Slot, ULit}
-import chisel3.internal.{
-  throwException,
-  AggregateViewBinding,
-  Builder,
-  ChildBinding,
-  InternalErrorException,
-  ViewBinding,
-  ViewParent
-}
+import chisel3.internal.{throwException, AggregateViewBinding, Builder, ChildBinding, ViewBinding, ViewParent}
 
 /** Represents lookup typeclass to determine how a value accessed from an original IsInstantiable
   *   should be tweaked to return the Instance's version
