+++ src/test/scala/chiselTests/ToTargetSpec.scala
@@ -3,7 +3,6 @@
 package chiselTests
 
 import chisel3._
-import chisel3.internal.ChiselException
 import circt.stage.ChiselStage
 
 class ToTargetSpec extends ChiselFlatSpec with Utils {
